fix(ci): checklist CC.8–10 gate failures (iOS lint, E2E.4, structure, routes)
Split oversized CourseChecklistSection for SwiftLint, add completed-feature
manifest entries for CC.8–10 companions, move assist SQL into course repo
for layering, and regenerate the route inventory golden for new assist routes.
