+++ clients/ios/Lextures/Features/Checklist/CourseChecklistSection.swift
@@ -10,21 +10,21 @@ struct CourseChecklistSection: View {
     let course: CourseSummary
     var initialFocus: String?
 
-    @State private var checklist: CourseChecklist?
-    @State private var loading = false
-    @State private var errorMessage: String?
-    @State private var rateLimitMessage: String?
+    @State private var controller: CourseChecklistController
     @State private var showCompleted = false
-    @State private var expandedCategories: Set<String> = []
     @State private var expandedItems: Set<String> = []
     @State private var dismissTarget: ChecklistItem?
-    @State private var actionError: String?
-    @State private var highlightAnchor: String?
-    @State private var highlightClearTask: Task<Void, Never>?
+
+    init(course: CourseSummary, initialFocus: String? = nil) {
+        self.course = course
+        self.initialFocus = initialFocus
+        _controller = State(initialValue: CourseChecklistController(courseCode: course.courseCode))
+    }
 
     private var isOnline: Bool { NetworkMonitor.shared.isOnline }
     private var summary: CourseChecklistSummary? {
-        checklist?.summary ?? CourseChecklistSummaryStore.shared.cached(courseCode: course.courseCode)
+        controller.checklist?.summary
+            ?? CourseChecklistSummaryStore.shared.cached(courseCode: course.courseCode)
     }
 
     var body: some View {
@@ -33,28 +33,35 @@ struct CourseChecklistSection: View {
 
             if !isOnline {
                 offlinePanel
-            } else if let errorMessage {
+            } else if let errorMessage = controller.errorMessage {
                 errorPanel(errorMessage)
-            } else if loading && checklist == nil {
+            } else if controller.loading && controller.checklist == nil {
                 LMSSkeletonList(count: 4)
-            } else if let checklist {
+            } else if let checklist = controller.checklist {
                 content(checklist)
             } else {
                 LMSSkeletonList(count: 3)
             }
 
-            if let rateLimitMessage {
+            if let rateLimitMessage = controller.rateLimitMessage {
                 Text(rateLimitMessage)
                     .font(.footnote)
                     .foregroundStyle(LexturesTheme.textSecondary(for: colorScheme))
             }
-            if let actionError {
+            if let actionError = controller.actionError {
                 Text(actionError)
                     .font(.footnote)
                     .foregroundStyle(LexturesTheme.coral)
             }
         }
-        .task { await loadFull() }
+        .task {
+            await controller.loadFull(
+                accessToken: session.accessToken,
+                isOnline: isOnline,
+                initialFocus: initialFocus,
+                reduceMotion: reduceMotion
+            )
+        }
         .sheet(item: $dismissTarget) { item in
             ChecklistDismissSheet(
                 item: item,
@@ -62,14 +69,20 @@ struct CourseChecklistSection: View {
                 onCancel: { dismissTarget = nil },
                 onConfirm: { reason, note in
                     dismissTarget = nil
-                    Task { await dismiss(item: item, reason: reason, note: note) }
+                    Task {
+                        await controller.dismiss(
+                            item: item,
+                            reason: reason,
+                            note: note,
+                            accessToken: session.accessToken,
+                            isOnline: isOnline
+                        )
+                    }
                 }
             )
             .presentationDetents([.medium, .large])
         }
     }
-
-    // MARK: - Header
 
     private var header: some View {
         VStack(alignment: .leading, spacing: 8) {
@@ -79,11 +92,16 @@ struct CourseChecklistSection: View {
                     .foregroundStyle(LexturesTheme.textPrimary(for: colorScheme))
                 Spacer()
                 Button {
-                    Task { await refresh() }
+                    Task {
+                        await controller.refresh(
+                            accessToken: session.accessToken,
+                            isOnline: isOnline
+                        )
+                    }
                 } label: {
                     Image(systemName: "arrow.clockwise")
                 }
-                .disabled(!isOnline || loading)
+                .disabled(!isOnline || controller.loading)
                 .accessibilityLabel(L.text("mobile.checklist.recheck"))
             }
 
@@ -93,10 +111,14 @@ struct CourseChecklistSection: View {
                     .font(.subheadline)
                     .foregroundStyle(LexturesTheme.textSecondary(for: colorScheme))
                     .accessibilityLabel(progress)
-                    .accessibilityValue("\(Int(CourseChecklistLogic.progressFraction(done: summary.done, total: summary.total) * 100)) percent")
+                    .accessibilityValue(
+                        "\(Int(CourseChecklistLogic.progressFraction(done: summary.done, total: summary.total) * 100)) percent"
+                    )
 
-                ProgressView(value: CourseChecklistLogic.progressFraction(done: summary.done, total: summary.total))
-                    .tint(LexturesTheme.accent(for: colorScheme))
+                ProgressView(
+                    value: CourseChecklistLogic.progressFraction(done: summary.done, total: summary.total)
+                )
+                .tint(LexturesTheme.accent(for: colorScheme))
             }
         }
     }
@@ -114,13 +136,19 @@ struct CourseChecklistSection: View {
             Text(message)
                 .foregroundStyle(LexturesTheme.coral)
             Button(L.text("mobile.checklist.retry")) {
-                Task { await loadFull(force: true) }
+                Task {
+                    await controller.loadFull(
+                        accessToken: session.accessToken,
+                        isOnline: isOnline,
+                        force: true,
+                        initialFocus: initialFocus,
+                        reduceMotion: reduceMotion
+                    )
+                }
             }
             .buttonStyle(.bordered)
         }
     }
-
-    // MARK: - Content
 
     @ViewBuilder
     private func content(_ checklist: CourseChecklist) -> some View {
@@ -132,7 +160,6 @@ struct CourseChecklistSection: View {
             Text(L.text("mobile.checklist.catalogEmpty"))
                 .foregroundStyle(LexturesTheme.textSecondary(for: colorScheme))
         } else {
-            // Offer show/hide whenever any items are done (not only when everything is done).
             if hasCompleted {
                 Toggle(isOn: $showCompleted) {
                     Text(showCompleted
@@ -147,7 +174,19 @@ struct CourseChecklistSection: View {
             }
 
             if !checklist.dismissed.isEmpty {
-                dismissedBlock(checklist.dismissed)
+                CourseChecklistDismissedBlock(
+                    items: checklist.dismissed,
+                    isOnline: isOnline,
+                    onRestore: { item in
+                        Task {
+                            await controller.restore(
+                                item: item,
+                                accessToken: session.accessToken,
+                                isOnline: isOnline
+                            )
+                        }
+                    }
+                )
             }
         }
     }
@@ -172,233 +211,63 @@ struct CourseChecklistSection: View {
     @ViewBuilder
     private func categoryBlock(_ category: ChecklistCategory) -> some View {
         let outstanding = CourseChecklistLogic.outstandingCount(in: category)
-        let expanded = expandedCategories.contains(category.id) || outstanding > 0
+        let expanded = controller.expandedCategories.contains(category.id) || outstanding > 0
         let items = CourseChecklistLogic.visibleItems(in: category, showCompleted: showCompleted)
 
         if items.isEmpty {
             EmptyView()
         } else {
-            categoryBlockBody(category, outstanding: outstanding, expanded: expanded, items: items)
-        }
-    }
-
-    private func categoryBlockBody(
-        _ category: ChecklistCategory,
-        outstanding: Int,
-        expanded: Bool,
-        items: [ChecklistItem]
-    ) -> some View {
-        VStack(alignment: .leading, spacing: 6) {
-            Button {
-                if expandedCategories.contains(category.id) {
-                    expandedCategories.remove(category.id)
-                } else {
-                    expandedCategories.insert(category.id)
-                }
-            } label: {
-                HStack {
-                    Image(systemName: expanded ? "chevron.down" : "chevron.right")
-                        .font(.caption.weight(.semibold))
-                    Text(category.title)
-                        .font(.subheadline.weight(.semibold))
-                        .foregroundStyle(LexturesTheme.textPrimary(for: colorScheme))
-                    Spacer()
-                    if outstanding > 0 {
-                        Text(L.format("mobile.checklist.outstandingCount", outstanding))
-                            .font(.caption)
-                            .foregroundStyle(LexturesTheme.textSecondary(for: colorScheme))
-                    }
-                }
-            }
-            .buttonStyle(.plain)
-            .accessibilityAddTraits(.isHeader)
-
-            if expanded {
-                ForEach(items) { item in
-                    itemRow(item)
-                }
-            }
-        }
-        .padding(.vertical, 4)
-    }
-
-    private func itemRow(_ item: ChecklistItem) -> some View {
-        let done = CourseChecklistLogic.isDone(item.status)
-        let evidenceCount = item.evidence?.rows.count ?? 0
-        let expanded = expandedItems.contains(item.id)
-
-        return VStack(alignment: .leading, spacing: 6) {
-            HStack(alignment: .top, spacing: 8) {
-                Image(systemName: done ? "checkmark.circle.fill" : statusIcon(item.status))
-                    .foregroundStyle(done ? LexturesTheme.brandTeal : LexturesTheme.textSecondary(for: colorScheme))
-                VStack(alignment: .leading, spacing: 4) {
-                    Text(item.title)
-                        .font(.body.weight(.medium))
-                        .strikethrough(done)
-                        .foregroundStyle(LexturesTheme.textPrimary(for: colorScheme))
-                    if let detail = item.detail, !detail.isEmpty {
-                        Text(detail)
-                            .font(.footnote)
-                            .foregroundStyle(LexturesTheme.textSecondary(for: colorScheme))
-                    }
-                    if !item.why.isEmpty {
-                        Text(item.why)
-                            .font(.caption)
-                            .foregroundStyle(LexturesTheme.textSecondary(for: colorScheme))
-                    }
-                    HStack(spacing: 6) {
-                        tierChip(item.tier)
-                        if let progress = item.progress {
-                            Text("\(progress.done) / \(progress.total)")
-                                .font(.caption2)
-                                .foregroundStyle(LexturesTheme.textSecondary(for: colorScheme))
-                        }
-                    }
-                }
-                Spacer(minLength: 0)
-                Menu {
-                    if !done {
-                        Button(L.text("mobile.checklist.open")) {
-                            openTarget(item.target)
-                        }
-                        Button(L.text("mobile.checklist.dismiss")) {
-                            dismissTarget = item
-                        }
-                        .disabled(!isOnline)
-                    }
-                    Button(L.text("mobile.checklist.recheckItem")) {
-                        Task { await recheck(item) }
-                    }
-                    .disabled(!isOnline)
-                } label: {
-                    Image(systemName: "ellipsis")
-                        .frame(minWidth: 44, minHeight: 44)
-                }
-                .accessibilityLabel(L.text("mobile.checklist.overflowMenu"))
-            }
-            .contentShape(Rectangle())
-            .onTapGesture {
-                if evidenceCount > 0 {
-                    toggleItem(item.id)
-                } else if !done {
-                    openTarget(item.target)
-                }
-            }
-            .accessibilityElement(children: .combine)
-            .accessibilityLabel(item.title)
-            .accessibilityValue(CourseChecklistLogic.accessibilityStatusValue(item.status))
-            .accessibilityHint(done ? "" : L.text("mobile.checklist.open"))
-
-            if evidenceCount > 0 {
+            VStack(alignment: .leading, spacing: 6) {
                 Button {
-                    toggleItem(item.id)
-                } label: {
-                    Text(expanded
-                          ? L.text("mobile.checklist.hideEvidence")
-                          : L.format("mobile.checklist.showEvidence", evidenceCount))
-                        .font(.footnote.weight(.semibold))
-                }
-                .buttonStyle(.plain)
-            }
-
-            if expanded, let evidence = item.evidence {
-                evidenceList(evidence)
-            }
-        }
-        .padding(10)
-        .background(LexturesTheme.cardBackground(for: colorScheme))
-        .clipShape(RoundedRectangle(cornerRadius: 12, style: .continuous))
-        .overlay(
-            RoundedRectangle(cornerRadius: 12, style: .continuous)
-                .stroke(
-                    highlightAnchor != nil && item.target?.anchor == highlightAnchor
-                        ? LexturesTheme.accent(for: colorScheme)
-                        : .clear,
-                    lineWidth: 2
-                )
-        )
-    }
-
-    private func evidenceList(_ evidence: ChecklistEvidence) -> some View {
-        VStack(alignment: .leading, spacing: 4) {
-            if let truncated = evidence.truncatedAt, truncated > 0, evidence.rows.count < truncated {
-                Text(L.format("mobile.checklist.evidenceTruncated", evidence.rows.count, truncated))
-                    .font(.caption2)
-                    .foregroundStyle(LexturesTheme.textSecondary(for: colorScheme))
-            }
-            ForEach(evidence.rows) { row in
-                Button {
-                    openTarget(row.target)
+                    if controller.expandedCategories.contains(category.id) {
+                        controller.expandedCategories.remove(category.id)
+                    } else {
+                        controller.expandedCategories.insert(category.id)
+                    }
                 } label: {
                     HStack {
-                        VStack(alignment: .leading, spacing: 2) {
-                            Text(row.label)
-                                .font(.subheadline)
-                                .foregroundStyle(LexturesTheme.textPrimary(for: colorScheme))
-                            if let sub = row.sublabel, !sub.isEmpty {
-                                Text(sub)
-                                    .font(.caption)
-                                    .foregroundStyle(LexturesTheme.textSecondary(for: colorScheme))
-                            }
-                        }
+                        Image(systemName: expanded ? "chevron.down" : "chevron.right")
+                            .font(.caption.weight(.semibold))
+                        Text(category.title)
+                            .font(.subheadline.weight(.semibold))
+                            .foregroundStyle(LexturesTheme.textPrimary(for: colorScheme))
                         Spacer()
-                        Image(systemName: "chevron.right")
-                            .font(.caption)
-                            .foregroundStyle(LexturesTheme.textSecondary(for: colorScheme))
-                    }
-                    .padding(.vertical, 6)
-                    .frame(minHeight: 44)
-                }
-                .buttonStyle(.plain)
-            }
-        }
-        .padding(.leading, 28)
-    }
-
-    private func dismissedBlock(_ items: [ChecklistItem]) -> some View {
-        DisclosureGroup {
-            ForEach(items) { item in
-                HStack {
-                    VStack(alignment: .leading, spacing: 2) {
-                        Text(item.title)
-                            .font(.subheadline)
-                        if let dismissal = item.dismissal {
-                            Text(L.format("mobile.checklist.dismissedBy", dismissal.byDisplayName, dismissal.reason))
-                                .font(.caption2)
+                        if outstanding > 0 {
+                            Text(L.format("mobile.checklist.outstandingCount", outstanding))
+                                .font(.caption)
                                 .foregroundStyle(LexturesTheme.textSecondary(for: colorScheme))
                         }
                     }
-                    Spacer()
-                    Button(L.text("mobile.checklist.restore")) {
-                        Task { await restore(item) }
-                    }
-                    .disabled(!isOnline)
-                    .font(.footnote.weight(.semibold))
                 }
-                .padding(.vertical, 6)
+                .buttonStyle(.plain)
+                .accessibilityAddTraits(.isHeader)
+
+                if expanded {
+                    ForEach(items) { item in
+                        CourseChecklistItemRow(
+                            item: item,
+                            isOnline: isOnline,
+                            expanded: expandedItems.contains(item.id),
+                            isHighlighted: controller.highlightAnchor != nil
+                                && item.target?.anchor == controller.highlightAnchor,
+                            onToggleEvidence: { toggleItem(item.id) },
+                            onOpen: { openTarget(item.target) },
+                            onDismiss: { dismissTarget = item },
+                            onRecheck: {
+                                Task {
+                                    await controller.recheck(
+                                        item: item,
+                                        accessToken: session.accessToken,
+                                        isOnline: isOnline
+                                    )
+                                }
+                            },
+                            onOpenEvidenceRow: { openTarget($0) }
+                        )
+                    }
+                }
             }
-        } label: {
-            Text(L.format("mobile.checklist.dismissedSection", items.count))
-                .font(.subheadline.weight(.semibold))
-        }
-    }
-
-    private func tierChip(_ tier: ChecklistTier) -> some View {
-        Text(tier == .essential
-              ? L.text("mobile.checklist.essentialTier")
-              : L.text("mobile.checklist.recommendedTier"))
-            .font(.caption2.weight(.semibold))
-            .padding(.horizontal, 6)
-            .padding(.vertical, 2)
-            .background(LexturesTheme.fieldBorder(for: colorScheme).opacity(0.35))
-            .clipShape(Capsule())
-    }
-
-    private func statusIcon(_ status: String) -> String {
-        switch CourseChecklistLogic.normalizeStatus(status) {
-        case .inProgress: return "circle.lefthalf.filled"
-        case .unknown: return "questionmark.circle"
-        default: return "circle"
+            .padding(.vertical, 4)
         }
     }
 
@@ -409,8 +278,6 @@ struct CourseChecklistSection: View {
             expandedItems.insert(id)
         }
     }
-
-    // MARK: - Navigation
 
     private func openTarget(_ target: ChecklistNavTarget?) {
         let resolved = CourseChecklistLogic.resolveTarget(
@@ -423,227 +290,27 @@ struct CourseChecklistSection: View {
             if let section = resolved.workspaceSection {
                 shell.activeCourseSection = section
                 if let anchor = resolved.focusAnchor {
-                    applyHighlight(anchor: anchor)
+                    controller.applyHighlight(anchor: anchor, reduceMotion: reduceMotion)
                 }
             }
         case .web:
             if let path = resolved.webPath {
-                let url = AppConfiguration.webURL(path: path)
-                LinkOpener.open(
-                    LinkOpener.Request(urlString: url.absoluteString, source: "checklist"),
-                    shell: shell
-                )
+                openWebPath(path)
             }
         case .unresolved:
             if let section = resolved.workspaceSection {
                 shell.activeCourseSection = section
             } else if let path = resolved.webPath {
-                let url = AppConfiguration.webURL(path: path)
-                LinkOpener.open(
-                    LinkOpener.Request(urlString: url.absoluteString, source: "checklist"),
-                    shell: shell
-                )
+                openWebPath(path)
             }
         }
     }
 
-    private func applyHighlight(anchor: String) {
-        highlightClearTask?.cancel()
-        highlightAnchor = anchor
-        let delay = reduceMotion ? 0.05 : CourseChecklistLogic.highlightDurationSeconds
-        highlightClearTask = Task {
-            try? await Task.sleep(nanoseconds: UInt64(delay * 1_000_000_000))
-            if !Task.isCancelled {
-                highlightAnchor = nil
-            }
-        }
-    }
-
-    // MARK: - Networking
-
-    private func loadFull(force: Bool = false) async {
-        guard isOnline else { return }
-        guard let token = session.accessToken else { return }
-        if !force, checklist != nil { return }
-        loading = true
-        errorMessage = nil
-        defer { loading = false }
-        do {
-            let result = try await LMSAPI.fetchCourseChecklist(
-                courseCode: course.courseCode,
-                accessToken: token
-            )
-            checklist = result
-            CourseChecklistSummaryStore.shared.applyChecklist(result)
-            for cat in result.categories where CourseChecklistLogic.outstandingCount(in: cat) > 0 {
-                expandedCategories.insert(cat.id)
-            }
-            if let focus = initialFocus {
-                applyHighlight(anchor: focus)
-            }
-        } catch let APIError.httpStatus(code, _) where code == 403 {
-            CourseChecklistSummaryStore.shared.markForbidden(courseCode: course.courseCode)
-            errorMessage = nil
-            checklist = nil
-        } catch {
-            errorMessage = Self.errorText(error)
-        }
-    }
-
-    private func refresh() async {
-        guard isOnline, let token = session.accessToken else { return }
-        loading = true
-        rateLimitMessage = nil
-        defer { loading = false }
-        do {
-            let result = try await LMSAPI.refreshCourseChecklist(
-                courseCode: course.courseCode,
-                accessToken: token
-            )
-            checklist = result
-            CourseChecklistSummaryStore.shared.applyChecklist(result)
-        } catch let APIError.httpStatus(code, _) where code == 429 {
-            rateLimitMessage = CourseChecklistLogic.rateLimitedMessage
-        } catch {
-            errorMessage = Self.errorText(error)
-        }
-    }
-
-    private func dismiss(item: ChecklistItem, reason: ChecklistDismissReason, note: String?) async {
-        guard isOnline, let token = session.accessToken, var current = checklist else { return }
-        // Optimistic move
-        let snapshot = current
-        removeItem(item.id, from: &current)
-        current.dismissed.insert(item, at: 0)
-        current.summary.dismissed += 1
-        if item.tier == .essential, CourseChecklistLogic.isOutstanding(item.status) {
-            current.summary.outstandingEssential = max(0, current.summary.outstandingEssential - 1)
-        }
-        if CourseChecklistLogic.isOutstanding(item.status) {
-            current.summary.outstandingTotal = max(0, current.summary.outstandingTotal - 1)
-        }
-        checklist = current
-        CourseChecklistSummaryStore.shared.applyChecklist(current)
-        do {
-            _ = try await LMSAPI.dismissChecklistItem(
-                courseCode: course.courseCode,
-                itemID: item.id,
-                reason: reason,
-                note: note,
-                accessToken: token
-            )
-            await loadFull(force: true)
-        } catch {
-            checklist = snapshot
-            CourseChecklistSummaryStore.shared.applyChecklist(snapshot)
-            actionError = Self.errorText(error)
-        }
-    }
-
-    private func restore(_ item: ChecklistItem) async {
-        guard isOnline, let token = session.accessToken else { return }
-        do {
-            _ = try await LMSAPI.restoreChecklistItem(
-                courseCode: course.courseCode,
-                itemID: item.id,
-                accessToken: token
-            )
-            await loadFull(force: true)
-        } catch {
-            actionError = Self.errorText(error)
-        }
-    }
-
-    private func recheck(_ item: ChecklistItem) async {
-        guard isOnline, let token = session.accessToken else { return }
-        do {
-            _ = try await LMSAPI.recheckChecklistItem(
-                courseCode: course.courseCode,
-                itemID: item.id,
-                accessToken: token
-            )
-            await loadFull(force: true)
-        } catch {
-            actionError = Self.errorText(error)
-        }
-    }
-
-    private static func errorText(_ error: Error) -> String {
-        (error as? LocalizedError)?.errorDescription ?? L.text("mobile.checklist.loadError")
-    }
-
-    private func removeItem(_ id: String, from checklist: inout CourseChecklist) {
-        checklist.categories = checklist.categories.map { cat in
-            var c = cat
-            c.items = c.items.filter { $0.id != id }
-            return c
-        }
-    }
-}
-
-// MARK: - Dismiss sheet
-
-private struct ChecklistDismissSheet: View {
-    let item: ChecklistItem
-    let isOnline: Bool
-    let onCancel: () -> Void
-    let onConfirm: (ChecklistDismissReason, String?) -> Void
-
-    @State private var reason: ChecklistDismissReason = .notApplicable
-    @State private var note = ""
-    @State private var showNote = false
-
-    var body: some View {
-        NavigationStack {
-            Form {
-                Section {
-                    Text(item.title)
-                        .font(.body.weight(.medium))
-                    Text(L.text("mobile.checklist.dismissDialogHelp"))
-                        .font(.footnote)
-                        .foregroundStyle(.secondary)
-                }
-                Section(L.text("mobile.checklist.dismissReasonLabel")) {
-                    Picker(L.text("mobile.checklist.dismissReasonLabel"), selection: $reason) {
-                        ForEach(ChecklistDismissReason.allCases) { r in
-                            Text(L.text(String.LocalizationValue(r.labelKey))).tag(r)
-                        }
-                    }
-                    .pickerStyle(.inline)
-                    .labelsHidden()
-                }
-                Section {
-                    if showNote {
-                        TextField(
-                            L.text("mobile.checklist.dismissNotePlaceholder"),
-                            text: $note,
-                            axis: .vertical
-                        )
-                        .lineLimit(3 ... 6)
-                    } else {
-                        Button(L.text("mobile.checklist.addNote")) {
-                            showNote = true
-                        }
-                    }
-                }
-                if !isOnline {
-                    Text(L.text("mobile.checklist.offlineMutations"))
-                        .foregroundStyle(.secondary)
-                }
-            }
-            .navigationTitle(L.text("mobile.checklist.dismissDialogTitle"))
-            .navigationBarTitleDisplayMode(.inline)
-            .toolbar {
-                ToolbarItem(placement: .cancellationAction) {
-                    Button(L.text("mobile.checklist.dismissCancel"), action: onCancel)
-                }
-                ToolbarItem(placement: .confirmationAction) {
-                    Button(L.text("mobile.checklist.dismissConfirm")) {
-                        onConfirm(reason, showNote ? note : nil)
-                    }
-                    .disabled(!isOnline)
-                }
-            }
-        }
+    private func openWebPath(_ path: String) {
+        let url = AppConfiguration.webURL(path: path)
+        LinkOpener.open(
+            LinkOpener.Request(urlString: url.absoluteString, source: "checklist"),
+            shell: shell
+        )
     }
 }
